@@ -8,15 +8,16 @@
     "start": "next start"
   },
   "dependencies": {
-    "cheerio": "^1.0.0-rc.5",
-    "github-markdown-css": "^4.0.0",
-    "next": "10.0.5",
-    "node-fetch": "^2.6.1",
+    "@vercel/analytics": "^1.0.1",
+    "cheerio": "^1.0.0-rc.12",
+    "eslint-config-next": "^13.4.3",
+    "github-markdown-css": "^5.2.0",
+    "next": "^13.4.3",
+    "node-fetch": "^3.3.1",
     "raw-loader": "^4.0.2",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-markdown": "^5.0.3",
-    "remark-html": "^13.0.2",
-    "request": "^2.88.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^8.0.7",
+    "rehype-raw": "^6.1.1"
   }
 }
